# core/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Stratt API
  description: >
    Estate-automation platform API: the typed estate graph (projection,
    charter §1.2), Views, and Run orchestration. Phase-0 surface.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
paths:
...
  /runs:
    get:
      operationId: listRuns
      summary: List recent Run summaries
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, format: int64, minimum: 1, maximum: 500, default: 100 }
      responses:
        "200":
          description: Recent Runs, newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Run" }
    post:
      operationId: startRun
      summary: Start a Run of the Phase-0 Workflow against a View
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StartRun" }
      responses:
        "201":
          description: The created Run
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Run" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    NotFound:
      description: The named object does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: The request is malformed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    StartRun:
      type: object
      description: >
        One Step: an Actuator against a View (charter §2.3: Actuator + content +
        params), or a targetless Connector Action (Action + params, ADR-0031).
        Set viewName+actuator OR action, not both.
      properties:
        viewName: { type: string }
        actuator:
          type: string
          description: >
            Names the Actuator as an OPAQUE routing key (ADR-0046 — the spine
            does not enumerate tools): any registered in-tree Actuator or plugin
            Actuator. An unknown name fails the Run terminally at launch. Matches
            the open-string workflow-step actuator field.
        action:
          type: string
          description: >
            A Connector Action (namespaced, e.g. awsec2/start). When set,
            this is a targetless typed operation — viewName/actuator are ignored
            and the CredentialRef `use` grant is the authz gate, not runner-on-View.
        dryRun:
          type: boolean
          description: Ask a DryRunnable Action to plan without side effects.
        params:
          type: object
          description: Actuator- or Action-interpreted Step params.
        slices:
          type: integer
          format: int64
          minimum: 1
          default: 1
          description: >
            Partition the target set across this many parallel execution
            Jobs (clamped to the target count).
        credentialRefs:
          type: array
          items: { type: string }
          description: >
            CredentialRef names to project into the execution pods. The
            launching Principal needs the `use` grant on each (checked at
            dispatch; use never implies read — charter §2.5).
        facetWriteScope:
          type: array
          items: { type: string }
          description: >
            Facet namespaces this Run may write back (ADR-0054). The effective
            allowlist enforced at the one governor is the actuator's registered
            grant ∩ this scope (least authority). Empty admits NO facet write-back
            (tight default); an entry not in the actuator's grant is rejected at
            authoring.
    Run:
      type: object
      required: [id, workflowId, status, startedAt]
      properties:
        id: { type: string }
        workflowId: { type: string }
        status:
          type: string
          enum: [pending, running, succeeded, failed, canceled, partial]
        viewRef: { type: string }
        viewVersion: { type: integer, format: int64 }
        triggeredBy:
          type: string
          description: >
            Name of the Trigger that fired this Run; absent for manual/API
            launches (charter §1.8 descent, Trigger → Run).
        baseline:
          type: string
          description: >
            Name of the Baseline whose cadence ran this check; absent for
            everything else (charter §1.8 descent, Baseline → Run).
        workflowRunId:
          type: string
          description: >
            The WorkflowRun this Run executed under, when it ran as a
            Workflow Step (charter §1.8 descent, Workflow → Run).
        stepName:
          type: string
          description: The Step within that Workflow.
        error:
          type: string
          description: >
            The terminal failure cause on a failed Run (charter §1.8) — the real
            message (e.g. a plugin Action's), so the descent shows WHY it failed.
            Absent unless the Run failed with a recorded cause.
        outputs:
          type: object
          description: >
            An Action Run's typed output values, validated against the Action's
            output Contract (charter §2.2, ADR-0031); absent for Actuator Runs.
        sites:
          type: array
          items: { type: string }
          description: >
            The execution loci this Run touched (charter §1.8 descent, ADR-0032):
            ["local"], a Site name, or a union when per-target routing fanned the
            Run across Sites. Absent for legacy Runs and targetless Actions.
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
...